move buf pkg to cosmosbuf
